+++ node_test.go
@@ -4,6 +4,7 @@ import (
 	"bytes"
 	"crypto/sha256"
 	"encoding/hex"
+	"math"
 	"math/rand"
 	"testing"
 
@@ -13,7 +14,7 @@ import (
 	iavlrand "github.com/cosmos/iavl/internal/rand"
 )
 
-func TestNode_encodedSize(t *testing.T) {
+func TestNode_maxEncodedSize(t *testing.T) {
 	node := &Node{
 		key:           iavlrand.RandBytes(10),
 		value:         iavlrand.RandBytes(10),
@@ -29,11 +30,11 @@ func TestNode_encodedSize(t *testing.T) {
 	}
 
 	// leaf node
-	require.Equal(t, 26, node.encodedSize())
+	require.Equal(t, 38, node.maxEncodedSize())
 
 	// non-leaf node
 	node.subtreeHeight = 1
-	require.Equal(t, 57, node.encodedSize())
+	require.Equal(t, 92, node.maxEncodedSize())
 }
 
 func TestNode_encode_decode(t *testing.T) {
@@ -42,37 +43,44 @@ func TestNode_encode_decode(t *testing.T) {
 		expectHex   string
 		expectError bool
 	}{
-		"nil":   {nil, "", true},
-		"empty": {&Node{}, "0000000000", false},
+		"nil":              {nil, "", true},
+		"version overflow": {&Node{version: math.MaxUint32 + 1}, "", true},
+		"empty":            {&Node{}, "000000000000", false},
 		"inner": {&Node{
 			subtreeHeight: 3,
 			version:       2,
 			size:          7,
 			key:           []byte("key"),
-			leftHash:      []byte{0x70, 0x80, 0x90, 0xa0},
-			rightHash:     []byte{0x10, 0x20, 0x30, 0x40},
-		}, "060e04036b657904708090a00410203040", false},
+			leftHash:      []byte{0x70, 0x80, 0x90, 0xa0, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00},
+			rightHash:     []byte{0x10, 0x20, 0x30, 0x40, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00},
+		}, "0300020700036b6579708090a0000000000000000000000000000000000000000000000000000000001020304000000000000000000000000000000000000000000000000000000000", false},
 		"leaf": {&Node{
 			subtreeHeight: 0,
 			version:       3,
 			size:          1,
 			key:           []byte("key"),
 			value:         []byte("value"),
-		}, "000206036b65790576616c7565", false},
+		}, "0000030100036b657976616c7565", false},
+		"large number": {&Node{
+			subtreeHeight: 0,
+			version:       math.MaxUint32,
+			size:          1 << 40,
+			key:           []byte("key"),
+			value:         []byte("value"),
+		}, "0013ffffffff000001036b657976616c7565", false},
 	}
 	for name, tc := range testcases {
 		tc := tc
 		t.Run(name, func(t *testing.T) {
-			var buf bytes.Buffer
-			err := tc.node.writeBytes(&buf)
+			bz, err := tc.node.Encode()
 			if tc.expectError {
 				require.Error(t, err)
 				return
 			}
 			require.NoError(t, err)
-			require.Equal(t, tc.expectHex, hex.EncodeToString(buf.Bytes()))
+			require.Equal(t, tc.expectHex, hex.EncodeToString(bz))
 
-			node, err := MakeNode(buf.Bytes())
+			node, err := MakeNode(bz)
 			require.NoError(t, err)
 			// since key and value is always decoded to []byte{} we augment the expected struct here
 			if tc.node.key == nil {
@@ -134,7 +142,9 @@ func TestNode_validate(t *testing.T) {
 	}
 }
 
-func BenchmarkNode_encodedSize(b *testing.B) {
+func BenchmarkNode_maxEncodedSize(b *testing.B) {
+	// deterministic test case is good for comparison
+	iavlrand.Seed(0)
 	node := &Node{
 		key:           iavlrand.RandBytes(25),
 		value:         iavlrand.RandBytes(100),
@@ -147,34 +157,84 @@ func BenchmarkNode_encodedSize(b *testing.B) {
 	b.ReportAllocs()
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
-		node.encodedSize()
+		_ = node.maxEncodedSize()
 	}
 }
 
-func BenchmarkNode_WriteBytes(b *testing.B) {
+func BenchmarkNode_Encode(b *testing.B) {
+	// deterministic test case is good for comparison
+	iavlrand.Seed(0)
 	node := &Node{
 		key:           iavlrand.RandBytes(25),
 		value:         iavlrand.RandBytes(100),
 		version:       rand.Int63n(10000000),
 		subtreeHeight: 1,
 		size:          rand.Int63n(10000000),
-		leftHash:      iavlrand.RandBytes(20),
-		rightHash:     iavlrand.RandBytes(20),
+		leftHash:      iavlrand.RandBytes(32),
+		rightHash:     iavlrand.RandBytes(32),
+	}
+	largeNode := &Node{
+		key:           iavlrand.RandBytes(25),
+		value:         iavlrand.RandBytes(100),
+		version:       rand.Int63n(10000000) + math.MaxUint16,
+		subtreeHeight: 1,
+		size:          rand.Int63n(10000000) + math.MaxUint32,
+		leftHash:      iavlrand.RandBytes(32),
+		rightHash:     iavlrand.RandBytes(32),
 	}
 	b.ResetTimer()
-	b.Run("NoPreAllocate", func(sub *testing.B) {
+	b.Run("small integer", func(sub *testing.B) {
 		sub.ReportAllocs()
 		for i := 0; i < sub.N; i++ {
-			var buf bytes.Buffer
-			_ = node.writeBytes(&buf)
+			_, _ = node.Encode()
 		}
 	})
-	b.Run("PreAllocate", func(sub *testing.B) {
+	b.Run("large integer", func(sub *testing.B) {
 		sub.ReportAllocs()
 		for i := 0; i < sub.N; i++ {
-			var buf bytes.Buffer
-			buf.Grow(node.encodedSize())
-			_ = node.writeBytes(&buf)
+			_, _ = largeNode.Encode()
+		}
+	})
+}
+
+func BenchmarkNode_Decode(b *testing.B) {
+	// deterministic test case is good for comparison
+	iavlrand.Seed(0)
+	node := &Node{
+		key:           iavlrand.RandBytes(25),
+		value:         iavlrand.RandBytes(100),
+		version:       rand.Int63n(10000000),
+		subtreeHeight: 1,
+		size:          rand.Int63n(10000000),
+		leftHash:      iavlrand.RandBytes(32),
+		rightHash:     iavlrand.RandBytes(32),
+	}
+	largeNode := &Node{
+		key:           iavlrand.RandBytes(25),
+		value:         iavlrand.RandBytes(100),
+		version:       rand.Int63n(10000000) + math.MaxUint16,
+		subtreeHeight: 1,
+		size:          rand.Int63n(10000000) + math.MaxUint32,
+		leftHash:      iavlrand.RandBytes(32),
+		rightHash:     iavlrand.RandBytes(32),
+	}
+
+	buf, err := node.Encode()
+	require.NoError(b, err)
+	largeNodeBuf, err := largeNode.Encode()
+	require.NoError(b, err)
+
+	b.ResetTimer()
+	b.Run("small number", func(sub *testing.B) {
+		sub.ReportAllocs()
+		for i := 0; i < sub.N; i++ {
+			_, _ = MakeNode(buf)
+		}
+	})
+	b.Run("large number", func(sub *testing.B) {
+		sub.ReportAllocs()
+		for i := 0; i < sub.N; i++ {
+			_, _ = MakeNode(largeNodeBuf)
 		}
 	})
 }
@@ -203,7 +263,7 @@ func BenchmarkNode_HashNode(b *testing.B) {
 		for i := 0; i < sub.N; i++ {
 			h := sha256.New()
 			buf := new(bytes.Buffer)
-			buf.Grow(node.encodedSize())
+			buf.Grow(node.maxEncodedSize())
 			require.NoError(b, node.writeHashBytes(buf))
 			_, err := h.Write(buf.Bytes())
 			require.NoError(b, err)